The quarantine data function was giving everything the same file name. Fixed this so the quarantined/excluded file type (e.g., fundamentals_daily) matches the data file. Affects: eod.py fundamentals.py fundamentals_daily.py
Changed the default threads to 8 since this is what most M1 and M2 Macs have.

When getting fundamentals daily data, checks the last updated time to avoid a call when there is no new data to get. This currently fails because the daily_last_updated field is incorrect, but if it gets fixed this will avoid a lot of calls.

Added signal handling because when doing multithreaded programming keyboard interrupt (ctrl-c) just doesn't work. Does a graceful exit.